# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Quick Automate Task Management API",
    "version": "1.0.0",
    "description": "RESTful CRUD Task Management API built using Node.js, Express.js and OpenAPI 3.0. Includes interactive Swagger documentation for testing API endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "Tasks",
      "description": "Operations related to task management"
    },
    {
      "name": "Health",
      "description": "API health monitoring"
    }
  ],
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Build AI Lead Generation Workflow"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "TaskInput": {
        "type": "object",
        "required": [
          "title",
          "done"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Create Client Onboarding Workflow"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Task created successfully."
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Task not found"
          }
        }
      }
    }
  },
  "paths": {
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get all tasks",
        "description": "Returns a list of all available tasks.",
        "responses": {
          "200": {
            "description": "List of tasks retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "title": "Build AI Lead Generation Workflow",
                    "done": true
                  },
                  {
                    "id": 2,
                    "title": "Integrate WhatsApp Cloud API",
                    "done": false
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a new task",
        "description": "Creates a new task and returns the created task.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              },
              "example": {
                "title": "Deploy Automation to Production",
                "done": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Task created successfully.",
                  "task": {
                    "id": 6,
                    "title": "Deploy Automation to Production",
                    "done": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Title and done fields are required."
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get task by ID",
        "description": "Returns a single task using its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                },
                "example": {
                  "id": 1,
                  "title": "Build AI Lead Generation Workflow",
                  "done": true
                }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Task not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Update an existing task",
        "description": "Updates an existing task using its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              },
              "example": {
                "title": "Deploy Automation to Production",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Task updated successfully.",
                  "task": {
                    "id": 4,
                    "title": "Deploy Automation to Production",
                    "done": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Title and done fields are required."
                }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Task not found"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete a task",
        "description": "Deletes a task using its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Task deleted successfully.",
                  "task": {
                    "id": 2,
                    "title": "Integrate WhatsApp Cloud API",
                    "done": false
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Task not found"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Check API health",
        "description": "Returns the current health status of the API.",
        "responses": {
          "200": {
            "description": "API is running successfully.",
            "content": {
              "application/json": {
                "example": {
                  "status": "OK",
                  "company": "Quick Automate",
                  "message": "Task Management API is running successfully."
                }
              }
            }
          }
        }
      }
    }
  }
}